fix auth

+++ src/main/java/ru/saynurdinov/moviefan/service/CollectionService.java
@@ -1,12 +1,12 @@
 package ru.saynurdinov.moviefan.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.saynurdinov.moviefan.DTO.CollectionDTO;
-import ru.saynurdinov.moviefan.DTO.CollectionEditDTO;
-import ru.saynurdinov.moviefan.DTO.CollectionPostDTO;
-import ru.saynurdinov.moviefan.DTO.CollectionPreviewDTO;
+import ru.saynurdinov.moviefan.DTO.*;
 import ru.saynurdinov.moviefan.mapper.PreviewMovieListMapper;
 import ru.saynurdinov.moviefan.model.Collection;
 import ru.saynurdinov.moviefan.model.Movie;
@@ -26,6 +26,7 @@ public class CollectionService {
     private final PreviewMovieListMapper movieListMapper;
     private final MovieRepository movieRepository;
     private final UserRepository userRepository;
+
     @Autowired
     public CollectionService(CollectionRepository collectionRepository, PreviewMovieListMapper movieListMapper, MovieRepository movieRepository, UserRepository userRepository) {
         this.collectionRepository = collectionRepository;
@@ -35,30 +36,38 @@ public class CollectionService {
     }
 
     @Transactional(readOnly = true)
-    public List<CollectionPreviewDTO> getAllByMovieIdNot(long userId, long movieId) {
-        List<Collection> collections = collectionRepository.findAllByOwner_IdAndMovies_IdNotOrMovies_IdIsNull(userId, movieId);
-        List<CollectionPreviewDTO> collectionPreviewDTOS = new ArrayList<>();
-        collections.forEach((collection) -> {
-            CollectionPreviewDTO collectionPreviewDTO =
-                    new CollectionPreviewDTO(collection.getId(), collection.getName());
-            collectionPreviewDTOS.add(collectionPreviewDTO);
-        });
-        return collectionPreviewDTOS;
+    public List<CollectionPreviewDTO> getAllByMovieIdNot(long movieId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication.isAuthenticated()) {
+            UserInfoDTO userInfo = (UserInfoDTO) authentication.getPrincipal();
+            List<Collection> collections = collectionRepository.findAllByOwner_IdAndMovies_IdNotOrMovies_IdIsNull(userInfo.getId(), movieId);
+            List<CollectionPreviewDTO> collectionPreviewDTOS = new ArrayList<>();
+            collections.forEach((collection) -> {
+                CollectionPreviewDTO collectionPreviewDTO =
+                        new CollectionPreviewDTO(collection.getId(), collection.getName());
+                collectionPreviewDTOS.add(collectionPreviewDTO);
+            });
+            return collectionPreviewDTOS;
+        } else throw new UsernameNotFoundException("Нет доступа. Неверные данные пользователя");
     }
 
     @Transactional(readOnly = true)
-    public List<CollectionDTO> getAllByUserId(long userId) {
-        List<Collection> collections = collectionRepository.findAllByOwner_Id(userId);
-        List<CollectionDTO> collectionDTOS = new ArrayList<>();
-        collections.forEach((collection -> {
-            CollectionDTO collectionDTO = new CollectionDTO();
-            collectionDTO.setId(collection.getId());
-            collectionDTO.setName(collection.getName());
-            collectionDTO.setOutline(collection.getOutline());
-            collectionDTO.setMoviesCollection(movieListMapper.toDTO(collection.getMovies()));
-            collectionDTOS.add(collectionDTO);
-        }));
-        return collectionDTOS;
+    public List<CollectionDTO> getAllByUserId() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication.isAuthenticated()) {
+            UserInfoDTO userInfo = (UserInfoDTO) authentication.getPrincipal();
+            List<Collection> collections = collectionRepository.findAllByOwner_Id(userInfo.getId());
+            List<CollectionDTO> collectionDTOS = new ArrayList<>();
+            collections.forEach((collection -> {
+                CollectionDTO collectionDTO = new CollectionDTO();
+                collectionDTO.setId(collection.getId());
+                collectionDTO.setName(collection.getName());
+                collectionDTO.setOutline(collection.getOutline());
+                collectionDTO.setMoviesCollection(movieListMapper.toDTO(collection.getMovies()));
+                collectionDTOS.add(collectionDTO);
+            }));
+            return collectionDTOS;
+        } else throw new UsernameNotFoundException("Нет доступа. Неверные данные пользователя");
     }
 
     @Transactional
@@ -68,10 +77,16 @@ public class CollectionService {
         if (movieOptional.isPresent() && collectionOptional.isPresent()) {
             Movie movie = movieOptional.get();
             Collection collection = collectionOptional.get();
-            collection.getMovies().add(movie);
-            movie.getCollections().add(collection);
-            collectionRepository.save(collection);
-            movieRepository.save(movie);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (authentication.isAuthenticated()) {
+                UserInfoDTO userInfo = (UserInfoDTO) authentication.getPrincipal();
+                if (userInfo.getId() == collection.getOwner().getId()) {
+                    collection.getMovies().add(movie);
+                    movie.getCollections().add(collection);
+                    collectionRepository.save(collection);
+                    movieRepository.save(movie);
+                } else throw new UsernameNotFoundException("Нет доступа. Неверные данные пользователя");
+            }
         }
     }
 
@@ -82,18 +97,25 @@ public class CollectionService {
         if (movieOptional.isPresent() && collectionOptional.isPresent()) {
             Movie movie = movieOptional.get();
             Collection collection = collectionOptional.get();
-            collection.getMovies().remove(movie);
-            movie.getCollections().remove(collection);
-            collectionRepository.save(collection);
-            movieRepository.save(movie);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (authentication.isAuthenticated()) {
+                UserInfoDTO userInfo = (UserInfoDTO) authentication.getPrincipal();
+                if (userInfo.getId() == collection.getOwner().getId()) {
+                    collection.getMovies().remove(movie);
+                    movie.getCollections().remove(collection);
+                    collectionRepository.save(collection);
+                    movieRepository.save(movie);
+                } else throw new UsernameNotFoundException("Нет доступа. Неверные данные пользователя");
+            }
         }
     }
 
     @Transactional
     public void createCollection(CollectionPostDTO collectionPostDTO) {
-        Optional<User> userOptional = userRepository.findById(collectionPostDTO.getUserId());
-        if (userOptional.isPresent()) {
-            User user = userOptional.get();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication.isAuthenticated()) {
+            UserInfoDTO userInfoDTO = (UserInfoDTO) authentication.getPrincipal();
+            User user = userRepository.findById(userInfoDTO.getId()).get();
             Collection collection = new Collection();
             collection.setName(collectionPostDTO.getName());
             if (!collectionPostDTO.getOutline().isEmpty()) {
@@ -103,7 +125,7 @@ public class CollectionService {
             collection.setOwner(user);
             collectionRepository.save(collection);
             userRepository.save(user);
-        }
+        } else throw new UsernameNotFoundException("Нет доступа. Неверные данные пользователя");
     }
 
     @Transactional
@@ -111,9 +133,15 @@ public class CollectionService {
         Optional<Collection> optionalCollection = collectionRepository.findById(collectionId);
         if (optionalCollection.isPresent()) {
             Collection collection = optionalCollection.get();
-            collection.setName(collectionEditDTO.getName());
-            collection.setOutline(collectionEditDTO.getOutline());
-            collectionRepository.save(collection);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (authentication.isAuthenticated()) {
+                UserInfoDTO userInfoDTO = (UserInfoDTO) authentication.getPrincipal();
+                if (collection.getOwner().getId() == userInfoDTO.getId()) {
+                    collection.setName(collectionEditDTO.getName());
+                    collection.setOutline(collectionEditDTO.getOutline());
+                    collectionRepository.save(collection);
+                } else throw new UsernameNotFoundException("Нет доступа. Неверные данные пользователя");
+            }
         }
     }
 
@@ -122,11 +150,17 @@ public class CollectionService {
         Optional<Collection> collectionOptional = collectionRepository.findById(collectionId);
         if (collectionOptional.isPresent()) {
             Collection collection = collectionOptional.get();
-            User owner = collection.getOwner();
-            owner.getCollections().remove(collection);
-            collection.getMovies().forEach(movie -> movie.getCollections().remove(collection));
-            collectionRepository.delete(collection);
-            userRepository.save(owner);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (authentication.isAuthenticated()) {
+                UserInfoDTO userInfoDTO = (UserInfoDTO) authentication.getPrincipal();
+                if (collection.getOwner().getId() == userInfoDTO.getId()) {
+                    User owner = collection.getOwner();
+                    owner.getCollections().remove(collection);
+                    collection.getMovies().forEach(movie -> movie.getCollections().remove(collection));
+                    collectionRepository.delete(collection);
+                    userRepository.save(owner);
+                } else throw new UsernameNotFoundException("Нет доступа. Неверные данные пользователя");
+            }
         }
     }
 }
