+++ src/main/java/ru/saynurdinov/moviefan/service/RatingService.java
@@ -1,10 +1,14 @@
 package ru.saynurdinov.moviefan.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.saynurdinov.moviefan.DTO.PostRateDTO;
 import ru.saynurdinov.moviefan.DTO.RatingDTO;
+import ru.saynurdinov.moviefan.DTO.UserInfoDTO;
 import ru.saynurdinov.moviefan.model.Movie;
 import ru.saynurdinov.moviefan.model.Rating;
 import ru.saynurdinov.moviefan.model.User;
@@ -32,30 +36,39 @@ public class RatingService {
     @Transactional
     public void rate(PostRateDTO postRateDTO) {
         Optional<Movie> movieOptional = movieRepository.findById(postRateDTO.getMovieId());
-        Optional<User> optionalUser = userRepository.findById(postRateDTO.getUserId());
-        if (movieOptional.isPresent() && optionalUser.isPresent()) {
+        if (movieOptional.isPresent()) {
             Movie movie = movieOptional.get();
-            User user = optionalUser.get();
-            Rating rating = new Rating();
-            rating.setOwner(user);
-            rating.setMovie(movie);
-            rating.setValue(postRateDTO.getValue());
-            user.getRatings().add(rating);
-            movie.getRatings().add(rating);
-            movie.calculateRating();
-            ratingRepository.save(rating);
-            userRepository.save(user);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (authentication.isAuthenticated()) {
+                UserInfoDTO userInfoDTO = (UserInfoDTO) authentication.getPrincipal();
+                User user = userRepository.findById(userInfoDTO.getId()).get();
+                Rating rating = new Rating();
+                rating.setOwner(user);
+                rating.setMovie(movie);
+                rating.setValue(postRateDTO.getValue());
+                user.getRatings().add(rating);
+                movie.getRatings().add(rating);
+                movie.calculateRating();
+                ratingRepository.save(rating);
+                userRepository.save(user);
+            }
+            else throw new UsernameNotFoundException("Нет доступа. Данные пользователя неверны");
         }
     }
 
     @Transactional(readOnly = true)
-    public RatingDTO get(long movieId, long userId) {
-        Optional<Rating> optionalRating = ratingRepository.findByMovie_IdAndOwner_Id(movieId, userId);
-        if (optionalRating.isPresent()) {
-            Rating rating = optionalRating.get();
-            return new RatingDTO(rating.getId(), rating.getValue());
+    public RatingDTO get(long movieId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication.isAuthenticated()) {
+            UserInfoDTO userInfoDTO = (UserInfoDTO) authentication.getPrincipal();
+            Optional<Rating> optionalRating = ratingRepository.findByMovie_IdAndOwner_Id(movieId, userInfoDTO.getId());
+            if (optionalRating.isPresent()) {
+                Rating rating = optionalRating.get();
+                return new RatingDTO(rating.getId(), rating.getValue());
+            }
+            return null;
         }
-        return null;
+        else throw new UsernameNotFoundException("Нет доступа. Данные пользователя неверны");
     }
 
     @Transactional
@@ -65,12 +78,19 @@ public class RatingService {
             Rating rating = optionalRating.get();
             Movie movie = rating.getMovie();
             User user = rating.getOwner();
-            movie.getRatings().remove(rating);
-            user.getRatings().remove(rating);
-            movie.calculateRating();
-            movieRepository.save(movie);
-            userRepository.save(user);
-            ratingRepository.delete(rating);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if (authentication.isAuthenticated()) {
+                UserInfoDTO userInfoDTO = (UserInfoDTO) authentication.getPrincipal();
+                if (userInfoDTO.getId() == user.getId()) {
+                    movie.getRatings().remove(rating);
+                    user.getRatings().remove(rating);
+                    movie.calculateRating();
+                    movieRepository.save(movie);
+                    userRepository.save(user);
+                    ratingRepository.delete(rating);
+                }
+                else throw new UsernameNotFoundException("Нет доступа. Данные пользователя неверны");
+            }
         }
     }
 }
